+++ backend/src/users/controllers/users/users.controller.ts
@@ -8,35 +8,52 @@ import {
   Param,
   Post,
   Put,
+  Req,
   Res,
+  UnauthorizedException,
+  UseGuards,
   UsePipes,
   ValidationPipe,
 } from '@nestjs/common';
+import { Request } from 'express';
+import { Types } from 'mongoose';
+import { AuthenticatedGuard } from '../../../auth/utils/LocalGuard';
+import { RequestCustom } from '../../../types/User';
 import { CreateUserDto } from '../../dtos/CreateUsers.dto';
 import { UpdateUserDto } from '../../dtos/UpdateUsers.dto';
+import { User } from '../../schema/user.schema';
 import { UsersService } from '../../services/users/users.service';
 
 @Controller('users')
 export class UsersController {
   constructor(private readonly userService: UsersService) {}
 
+  @UseGuards(AuthenticatedGuard)
   @Post()
   @UsePipes(ValidationPipe)
-  async createUser(@Res() response, @Body() user: CreateUserDto) {
-    try {
-      const newUser = await this.userService.create(user);
-      return response.status(HttpStatus.CREATED).json({
-        newUser,
-      });
-    } catch (err) {
-      return response.status(HttpStatus.BAD_REQUEST).json({
-        statusCode: 400,
-        message: 'Error: User not created',
-        error: 'Bad Request',
-      });
+  async createUser(
+    @Res() response,
+    @Body() user: CreateUserDto,
+    @Req() request: RequestCustom,
+  ) {
+    if (request.user.role === 'admin') {
+      try {
+        const newUser = await this.userService.create(user);
+        return response.status(HttpStatus.CREATED).json({
+          newUser,
+        });
+      } catch (err) {
+        return response.status(HttpStatus.BAD_REQUEST).json({
+          statusCode: 400,
+          message: 'Error: User not created',
+          error: 'Bad Request',
+        });
+      }
     }
+    throw new UnauthorizedException();
   }
 
+  @UseGuards(AuthenticatedGuard)
   @Get(':id')
   async getUser(@Res() response, @Param('id') userId: string) {
     try {
@@ -53,6 +70,7 @@ export class UsersController {
     }
   }
 
+  @UseGuards(AuthenticatedGuard)
   @Get()
   async getUsers(@Res() response) {
     try {
@@ -68,40 +86,53 @@ export class UsersController {
       return response.status(err.status).json(err.response);
     }
   }
-
+  @UseGuards(AuthenticatedGuard)
   @Put(':id')
   async updateUser(
     @Res() response,
     @Param('id') userId: string,
     @Body() updateUserDto: UpdateUserDto,
+    @Req() request: RequestCustom,
   ) {
-    try {
-      const updatedUser = await this.userService.update(updateUserDto, userId);
-      if (!updatedUser) {
-        throw new NotFoundException(`User #${userId} not found`);
+    if (request.user.role === 'admin') {
+      try {
+        const updatedUser = await this.userService.update(
+          updateUserDto,
+          userId,
+        );
+        if (!updatedUser) {
+          throw new NotFoundException(`User #${userId} not found`);
+        }
+        return response.status(HttpStatus.OK).json({
+          message: 'User data updated successfully',
+          updatedUser,
+        });
+      } catch (err) {
+        return response.status(err.status).json(err.response);
       }
-      return response.status(HttpStatus.OK).json({
-        message: 'User data updated successfully',
-        updatedUser,
-      });
-    } catch (err) {
-      return response.status(err.status).json(err.response);
     }
   }
 
+  @UseGuards(AuthenticatedGuard)
   @Delete(':id')
-  async deleteCustomer(@Res() response, @Param('id') userId: string) {
-    try {
-      const deletedUser = await this.userService.delete(userId);
-      if (!deletedUser) {
-        throw new NotFoundException(`User #${userId} not found`);
+  async deleteCustomer(
+    @Res() response,
+    @Param('id') userId: string,
+    @Req() request: RequestCustom,
+  ) {
+    if (request.user.role === 'admin') {
+      try {
+        const deletedUser = await this.userService.delete(userId);
+        if (!deletedUser) {
+          throw new NotFoundException(`User #${userId} not found`);
+        }
+        return response.status(HttpStatus.OK).json({
+          message: 'User data deleted successfully',
+          deletedUser,
+        });
+      } catch (err) {
+        return response.status(err.status).json(err.response);
       }
-      return response.status(HttpStatus.OK).json({
-        message: 'User data deleted successfully',
-        deletedUser,
-      });
-    } catch (err) {
-      return response.status(err.status).json(err.response);
     }
   }
 }
